Add initial libev support from chapel.

+++ c/tcp_server_main.c
@@ -1,8 +1,8 @@
-#include <netinet/in.h>
+#include <sys/socket.h>
 #include <stdlib.h>
 #include <tcp_server.h>
 
-void handle_received_data(int fd, char *buffer, int read, int buffer_size) {
+void handle_received_data(int fd, char *buffer, size_t read, size_t buffer_size) {
   plog("message:%s", buffer);
   if (read >= 2 && buffer[0] == ':' && buffer[1] == 'q') {
     plog("quitting\n");
